build: add component Form
